requirements.txt
app==0.0.1
APScheduler==3.6.3
asn1crypto==1.2.0
bilibili-api==3.1.3
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
click==7.1.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
Django==2.1.7
django-ckeditor==5.8.0
django-cors-headers==3.2.0
django-filter==2.2.0
django-js-asset==1.2.2
django-simpleui==3.9
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.5.0
drf-yasg==1.17.0
esdk-obs-python==3.20.11
Flask==1.1.2
icalendar==4.0.9
idna==2.8
inflection==0.3.1
itsdangerous==1.1.0
itypes==1.1.0
Jinja2==2.10.3
lxml==4.5.2
Markdown==3.1.1
MarkupSafe==1.1.1
mysqlclient==1.4.6
packaging==19.2
Pillow==9.4.0
pycparser==2.19
pycrypto==2.6.1
pycryptodomex==3.9.4
PyJWT==1.7.1
PyMySQL==0.9.3
pyOpenSSL==19.1.0
pyparsing==2.4.5
python-alipay-sdk==2.0.1
pytz==2019.3
PyYAML==5.3.1
redis==3.5.3
requests==2.24.0
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.2.0
Serializer==0.2.1
simplejson==3.17.0
six==1.13.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.7
Werkzeug==1.0.1
wget==3.2
